requirements.txt
streamlit>=1.32.0
openai>=1.14.0
anthropic>=0.20.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
